splitting out the helpers from the functionality.
