+++ src/helpers/packet_helpers.go
@@ -1,4 +1,4 @@
-package parsers
+package helpers
 
 import (
 	"bytes"
